Added search function and other pages
added search function on index have not implemented it. added links to edit and delete members. the search member details should show on search page. have left placeholders for most of the stuff please change them to the searched member details. also added a sales page as a place holder.